+++ packages/viz/src/index.js
@@ -43,7 +43,12 @@ import DataUtil from './utils/data'
 import getParametersChanges from './utils/parametersHistory'
 import createPrintPDFPackage from './modules/print'
 import { generatePDFStats, generatePumpSettings, selectDailyViewData } from './utils/print/data'
-import { delayShowCbgTracesOnFocus, unfocusTrendsCbgSlice } from './redux/actions/index'
+import {
+  delayShowCbgTracesOnFocus,
+  focusTrendsCbgDateTrace,
+  unfocusTrendsCbgDateTrace,
+  unfocusTrendsCbgSlice
+} from './redux/actions/index'
 
 const components = {
   FocusedRangeLabels,
@@ -100,5 +105,7 @@ export {
   getLongDayHourFormat,
   formatParameterValue,
   delayShowCbgTracesOnFocus,
+  focusTrendsCbgDateTrace,
+  unfocusTrendsCbgDateTrace,
   unfocusTrendsCbgSlice
 }