+++ CHANGELOG.md
@@ -4,6 +4,7 @@ It is based on Tidepool Blip 1.27.
 
 ## 3.2.0-beta - 2023-02-01
 ### Engineering use
+- YLP-2044 Remove redux settings store
 - YLP-2065 Migrate CBGDateTraceLabel from viz to dumb
 - YLP-2068 Update minor dependencies
 - YLP-2069 Migrate CBGSlicesContainer from viz to dumb