+++ CHANGELOG.md
@@ -16,6 +16,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2096 Migrate CBG date traces animated to dumb
 - YLP-2097 Migrate RangeSelect from viz to dumb
 - YLP-2110 Update i18next from 19.9.2 to 22.4.9
+- YLP-2118 Update axios and minor dependencies
 
 ## 3.1.0 - 2023-01-31
 ### Fixed