+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.1.0-rc1 - 2023-01-31
+## 3.1.0-rc2 - 2023-02-07
 ### Fixed
 - YLP-1682 Labels in PDF reports are not clear
 - YLP-1704 PDF Report Generation failed